ULTIMATE UPDATE EXTRAVAGANZA
EVERYTHING YOU WOULD EVER NEED IS HERE! Ashley's Item generation works
but they are too numerous and I believe her random placement affects the
run time speed as start up is really slow. Not a big problem but the
number of items really needs to be toned down. Also, they only load on
the first load of the map. So, I recommend you modify the readLevel
function with your items. Maybe push all of your item vectors onto
another vector to do it? Just a thought. Also at this point, the
character only walks over the items now. Please everyone grab these
files so we can all be on the same page. Alex, can you also had some
default values to the constructor of a character as I am too lazy and
forgot to do so. Stuff like "health = 1" "Armor = 0" and "Level = 1".
That is all.